Stop modal from closing if there is an error
When there is an error, a notification is fired, but the modal closes. This means that to rectify any issues the user has to open the modal again, upload the file and fill out the form (if it has been defined).

Now, the action is halted if there is an error so that no data is lost in the modal form.